+++ packages/implementations/react/src/atomic/Progress/src/components/Circle/circle.definitions.ts
@@ -1,0 +1,1 @@
+export const circunference = 67 * Math.PI * 2;