+++ packages/implementations/react/src/atomic/Progress/src/components/Circle/circle.types.ts
@@ -1,0 +1,6 @@
+import { HTMLAttributes } from "react";
+import { ProgressVariants } from "@rarui/styles";
+import { ProgressProperties } from "../../progress.types";
+
+export type CircleProperties = ProgressProperties & ProgressVariants;
+export type CircleProps = CircleProperties & HTMLAttributes<HTMLDivElement>;