+++ packages/implementations/react/src/atomic/Progress/src/index.ts
@@ -1,0 +1,6 @@
+import { Progress } from "./Progress";
+
+export type { ProgressProps } from "./progress.types";
+export { Progress } from "./Progress";
+
+export default Progress;