+++ packages/styles/src/packages/atomic/progress/index.ts
@@ -1,0 +1,8 @@
+import { progressSprinkle, progressStyles } from "./rarui-progress.css";
+
+export const progress = {
+  classnames: progressStyles,
+  sprinkle: progressSprinkle,
+};
+
+export type { ProgressVariants } from "./rarui-progress.types";