+++ packages/styles/src/packages/atomic/progress/rarui-progress.types.ts
@@ -1,0 +1,4 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { progressStyles } from "./rarui-progress.css";
+
+export type ProgressVariants = RecipeVariants<typeof progressStyles.circle>;